Moving files and creating generateGame function file
