+++ sudokuru/app/Components/Sudoku Board/SudokuBoard.tsx
@@ -21,10 +21,11 @@ import {
   getGroupsFromHint,
   getNumberOfGroupsAssignedForNumber, getPlacementsFromHint, getRemovalsFromHint, parseApiAndAddNotes, replaceChar,
   strPuzzleToArray, updateBoardWithNumber, getHint
-} from "./BoardFunctions";
+} from "./Functions/BoardFunctions";
 import Cell from "./Components/Cell";
 import ActionRow from "./Components/ActionRow";
 import HintSection from "./Components/HintSection";
+import {generateGame} from "./Functions/generateGame";
 
 // Sudokuru Package Import
 const sudokuru = require("../../../node_modules/sudokuru/dist/bundle.js");
@@ -92,97 +93,6 @@ const styles = (cellSize, sizeConst, theme) => StyleSheet.create({
     color: '#FFFFFF',
   },
 });
-
-async function generateGame(url, props) {
-
-  let token = null;
-
-  await getKeyString("access_token").then(result => {
-    token = result;
-  });
-
-  let gameData = null;
-
-  if (props.gameType == "StartGame"){
-    gameData = await Puzzles.startGame(url, props.difficulty, props.strategies, token).then(
-        game => {
-          // If game object is not returned, you get redirected to Main Page
-          if (game == null){
-            //navigation.navigate("Home");
-            return;
-          }
-          let board = makeBoard(strPuzzleToArray(game[0].puzzle), game[0].puzzle);
-          return {
-            board,
-            history: List.of(board),
-            historyOffSet: 0,
-            solution: game[0].puzzleSolution,
-            activeGame: game,
-          };
-        }
-    );
-  }
-  else if (props.gameType == "ResumeGame"){
-    gameData = await Puzzles.getGame(url, token).then(
-        game => {
-          // If game object is not returned, you get redirected to Main Page
-          if (game == null){
-            //navigation.navigate("Home");
-            return;
-          }
-          let board = makeBoard(strPuzzleToArray(game[0].moves[game[0].moves.length-1].puzzleCurrentState), game[0].puzzle);
-          board = parseApiAndAddNotes(board, game[0].moves[game[0].moves.length-1].puzzleCurrentNotesState, false);
-          return {
-            board,
-            history: List.of(board),
-            historyOffSet: 0,
-            solution: game[0].puzzleSolution,
-            activeGame: game,
-          };
-        }
-    );
-  }
-  else if (props.gameType == 'StartDrill'){
-    let token = null;
-    await getKeyString("access_token").then(
-        result => {
-          token = result;
-        });
-
-    let { board, originalBoard, puzzleSolution } = await Drills.getGame(url, props.strategies, token).then(game => {
-      // null check to verify that game is loaded in.
-      if (game == null){
-        //navigation.navigate("Home");
-        return;
-      }
-      let board = makeBoard(strPuzzleToArray(game.puzzleCurrentState), game.puzzleCurrentState);
-      board = parseApiAndAddNotes(board, game.puzzleCurrentNotesState, true);
-      let originalBoard = makeBoard(strPuzzleToArray(game.puzzleCurrentState), game.puzzleCurrentState);
-      originalBoard = parseApiAndAddNotes(originalBoard, game.puzzleCurrentNotesState, true);
-      let puzzleSolution = game.puzzleSolution;
-      return { board, originalBoard, puzzleSolution };
-    });
-
-    let drillSolutionCells = getDrillSolutionCells(board, puzzleSolution, props.strategies);
-
-    return {
-      board, history: List.of(board), historyOffSet: 0, drillSolutionCells, originalBoard, solution: puzzleSolution
-    };
-  }
-  else if (props.gameType == 'Demo'){
-    game = Puzzles.getRandomGame()
-    let board = makeBoard(strPuzzleToArray(game[0].puzzle), game[0].puzzle);
-    return {
-      board,
-      history: List.of(board),
-      historyOffSet: 0,
-      solution: game[0].puzzleSolution,
-      activeGame: game,
-    };
-  }
-
-  return gameData;
-}
 
 const Puzzle = (props) => {
   const { board, renderCell } = props;
